requirements.txt
flask
pillow
pdf2image
poppler-utils  # Required for pdf2image
